+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -3,13 +3,13 @@
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.22.4
+-- mp7_ugt (=FRAME_VERSION): v1.24.0
 
 -- use "GT_VERSION" as mp7_ugt release fw version (used for tag name).
--- gt: v1.22.4
--- frame: v1.4.0 (see frame.vhd)
--- gtl: v1.18.2 (see gtl_module_tpl.vhd)
--- fdl: v1.4.1 (see fdl_module.vhd)
+-- gt: v1.24.0
+-- frame: v1.4.1 (see frame.vhd)
+-- gtl: v1.12.0 (see gtl_module_tpl.vhd)
+-- mp7_ugt (=FRAME_VERSION): v1.23.2
 
 -- HB 2023-04-08: v1.22.2 - Changed in scripts/mp7_patch.py (for area_constraints.tcl).
 -- HB 2023-04-07: v1.22.1 - Used Tx links 28-31 for scouting (frame/output_mux.vhd) without links 32..71.
@@ -34,7 +34,9 @@
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
 -- gtl history:
--- HB 2023-04-27: v1.18.2: Implemented topological trigger.
+-- HB 2023-09-05: v1.20.0: Implemented topological trigger.
+-- HB 2023-09-01: v1.19.1: Added zdc_condition.vhd (was missing) and updated sim and dep file.
+-- HB 2023-08-25: v1.19.0: Used link 71 as ZDC 5G input. ZDC data structure changed (16 bits only on 5G link). Updated top_decl.vhd.
 -- HB 2023-04-15: v1.18.1: Bug fix muon index bits.
 -- HB 2023-03-14: v1.18.0: Implemented "MUS2" (new hadronic shower bit). Module for "anomaly detection trigger (ADT)". Added ZDC 10G optical on link 11.
 -- HB 2023-02-14: v1.17.7: Inserted cut for muon index bits.
@@ -92,8 +94,8 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- GT firmware version
     constant GT_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GT_MINOR_VERSION      : integer range 0 to 255 := 22;
-    constant GT_REV_VERSION        : integer range 0 to 255 := 4;
+    constant GT_MINOR_VERSION      : integer range 0 to 255 := 24;
+    constant GT_REV_VERSION        : integer range 0 to 255 := 0;
 	constant GT_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(GT_MAJOR_VERSION, 8)) &
            std_logic_vector(to_unsigned(GT_MINOR_VERSION, 8)) &
@@ -101,11 +103,11 @@ package gt_mp7_core_pkg is
 -- FRAME firmware version
     constant FRAME_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 4;
-    constant FRAME_REV_VERSION        : integer range 0 to 255 := 0;
+    constant FRAME_REV_VERSION        : integer range 0 to 255 := 1;
 -- GTL firmware (fix part) version
     constant GTL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GTL_FW_MINOR_VERSION      : integer range 0 to 255 := 18;
-    constant GTL_FW_REV_VERSION        : integer range 0 to 255 := 2;
+    constant GTL_FW_MINOR_VERSION      : integer range 0 to 255 := 20;
+    constant GTL_FW_REV_VERSION        : integer range 0 to 255 := 0;
 -- FDL firmware version
     constant FDL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant FDL_FW_MINOR_VERSION      : integer range 0 to 255 := 4;