+++ firmware/hdl/packages/gtl_pkg.vhd
@@ -2,6 +2,8 @@
 -- Package for constant and type definitions of GTL firmware in Global Trigger Upgrade system.
 
 -- Version history:
+-- HB 2023-07-28: bug fixed "type zdc_array ...".
+-- HB 2023-07-25: new ZDC data structure.
 -- HB 2023-03-06: added hadronic shower trigger bit MUS2.
 -- HB 2023-03-01: updated constants for ZDC.
 -- HB 2022-10-17: added constant NR_INPUT_LANES (for reduced NR_LANES in frame.vhd).
@@ -387,16 +389,29 @@ constant MBT1HFM_COUNT_HIGH : natural := 3;
 -- *******************************************************************************************************
 -- HB 2022-10-10: inserted ZDC
 constant NR_ZDC_OBJECTS : positive := 6;
-constant EN_MINUS_BIT_LOW : natural := 16; -- EN_MINUS (ZDC-) on frame 0 of ZDC link
-constant EN_MINUS_BIT_HIGH : natural := 25;
-constant EN_PLUS_BIT_LOW : natural := 0; -- EN_PLUS (ZDC+) on frame 1 of ZDC link
+
+-- Email Jeremy Mans 24.7.2023:
+-- Definition of ZDC data (lower 16 bits on 5G link frames):
+
+-- [8 bits of zero] [3c or 7c]
+-- [6 bits of zero] [10 bit energy from ZDC-]
+-- [6 bits of zero] [10 bit energy from ZDC+]
+-- [16 bits of zero]
+-- [4 bits of zero] [12 bits of counter]
+-- [16 bits of zero]
+
+constant EN_MINUS_FRAME : natural := 1; -- EN_MINUS (ZDC-) on frame 1 of ZDC link
+constant EN_MINUS_BIT_LOW : natural := 0;
+constant EN_MINUS_BIT_HIGH : natural := 9;
+constant EN_PLUS_FRAME : natural := 2; -- EN_PLUS (ZDC+) on frame 2 of ZDC link
+constant EN_PLUS_BIT_LOW : natural := 0;
 constant EN_PLUS_BIT_HIGH : natural := 9;
 constant ZDC_BIT_LOW : natural := 0;
 constant ZDC_BIT_HIGH : natural := 9;
-constant MAX_ZDC_BITS : natural := 32;
+constant MAX_ZDC_BITS : natural := 16;
 constant ZDC_THR_BITS : natural := 16;
-type zdc_5g_array is array (0 to LINK_FRAMES-1) of std_logic_vector(SW_DATA_WIDTH-1 downto 0);
-type zdc_array is array (0 to NR_ZDC_OBJECTS-1) of std_logic_vector(ZDC_BIT_HIGH-ZDC_BIT_LOW downto 0);
+-- type zdc_5g_array is array (0 to LINK_FRAMES-1) of std_logic_vector(SW_DATA_WIDTH-1 downto 0);
+type zdc_array is array (0 to NR_ZDC_OBJECTS-1) of std_logic_vector(SW_DATA_WIDTH-1 downto 0);
 type bx_zdc_array is array (0 to BX_PIPELINE_STAGES-1) of std_logic_vector(ZDC_BIT_HIGH-ZDC_BIT_LOW downto 0);
 
 -- *******************************************************************************************************
@@ -478,8 +493,8 @@ type bx_data_record is record
     cent7 : bx_cent_array;
     ext_cond : bx_ext_cond_array;
     mus0, mus1, mus2, musoot0, musoot1 : mus_bit_array;
-    zdcp : bx_zdc_array;
     zdcm : bx_zdc_array;
+    zdcp : bx_zdc_array;
 end record bx_data_record;
 
 -- ==== Correlations - begin ============================================================