+++ firmware/hdl/packages/lhc_data_pkg.vhd
@@ -2,6 +2,7 @@
 -- Package for definitions of lhc data.
 
 -- Version history:
+-- HB 2023-08-25: ZDC 5G on link 71.
 -- HB 2023-02-09: updated comments.
 -- HB 2022-10-08: zdc10g on link 11.
 -- HB 2022-09-08: cleaned up.
@@ -24,7 +25,7 @@
 -- tau                 6 (5..0)        1        8
 -- tau                 6 (11..6)       1        9
 -- esums               6               1       10
--- zdc                 6               1       11
+-- free                6               1       11
 -- ext-cond(63..0)     2               1       12
 -- ext-cond(127..64)   2               1       13
 -- ext-cond(191..128)  2               1       14
@@ -34,6 +35,7 @@
 -- muon                2 (3..2)        1        1
 -- muon                2 (5..4)        1        2
 -- muon                2 (7..6)        1        3
+-- zdc                 6               1       71
 -- __________________________________________________________
 -- Summary             72 (32 bits)    16
 --
@@ -46,7 +48,7 @@
 -- tau         5      4     3    2    1      0
 -- tau         11     10    9    8    7      6
 -- esums       HTmHF  ETmHF HTm* ETm* HT,TC* ET,ETTEM*
--- zdc (10G)   5      4     3    2    1      0     
+-- free        5      4     3    2    1      0     
 -- ext-cond    x      x     x    x    1      0
 -- ext-cond    x      x     x    x    3      2
 -- ext-cond    x      x     x    x    5      4
@@ -55,8 +57,9 @@
 -- muon        3h     3l    2h   2l   x      x
 -- muon        5h     5l    4h   4l   x      x
 -- muon        7h     7l    6h   6l   x      x
+-- zdc (5G)    5      4     3    2    1      0     
 --
--- *) in addition 4 MSBs for Minimum Bias bits. TC means "TOWERCOUNT" (ECAL sum)
+-- *) in addition 4 MSBs for Minimum Bias bits. TC means "TOWERCNT" (ECAL sum)
 
 --===============================================================================================--
 
@@ -74,11 +77,9 @@ package lhc_data_pkg is
     constant OFFSET_TAU_LANES : natural := 8;
     constant OFFSET_ESUMS_LANES : natural := 10;
 -- HB 2106-05-31: proposal for memory structure with all frames of calo links for extended test-vector-file structure (see lhc_data_pkg_all_frames.vhd)
-    constant OFFSET_ZDC10G_LANES : natural := 11;
+    constant OFFSET_LINK_11_LANES : natural := 11;
     constant OFFSET_EXT_COND_LANES : natural := 12;
-
--- HB 2022-10-10: lane number of ZDC 5G optical input
-    constant ZDC5G_LANE_NR : natural := 71;
+    constant OFFSET_ZDC5G_LANES : natural := 71;
 
 --  for simspy memory (test with ipb_dpmem_4096_32)
     constant SW_DATA_WIDTH : integer := 32;
@@ -97,13 +98,19 @@ package lhc_data_pkg is
     constant HTM_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant ETMHF_DATA_WIDTH : integer := SW_DATA_WIDTH;
     constant HTMHF_DATA_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_0_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_1_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_2_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_3_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_4_WIDTH : integer := SW_DATA_WIDTH;
-    constant ZDC10G_5_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_0_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_1_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_2_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_3_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_4_WIDTH : integer := SW_DATA_WIDTH;
+    constant LINK_11_FR_5_WIDTH : integer := SW_DATA_WIDTH;
     constant EXTERNAL_CONDITIONS_DATA_WIDTH : integer := SW_DATA_WIDTH*8;
+    constant ZDC5G_0_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_1_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_2_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_3_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_4_WIDTH : integer := SW_DATA_WIDTH;
+    constant ZDC5G_5_WIDTH : integer := SW_DATA_WIDTH;
 
     constant LHC_DATA_WIDTH : integer :=
         (
@@ -117,10 +124,13 @@ package lhc_data_pkg is
             HTM_DATA_WIDTH +
             ETMHF_DATA_WIDTH +
             HTMHF_DATA_WIDTH +
-            ZDC10G_0_WIDTH + ZDC10G_1_WIDTH +
-            ZDC10G_2_WIDTH + ZDC10G_3_WIDTH +
-            ZDC10G_4_WIDTH + ZDC10G_5_WIDTH +
-            EXTERNAL_CONDITIONS_DATA_WIDTH
+            LINK_11_FR_0_WIDTH + LINK_11_FR_1_WIDTH +
+            LINK_11_FR_2_WIDTH + LINK_11_FR_3_WIDTH +
+            LINK_11_FR_4_WIDTH + LINK_11_FR_5_WIDTH +
+            EXTERNAL_CONDITIONS_DATA_WIDTH +
+            ZDC5G_0_WIDTH + ZDC5G_1_WIDTH +
+            ZDC5G_2_WIDTH + ZDC5G_3_WIDTH +
+            ZDC5G_4_WIDTH + ZDC5G_5_WIDTH
         );
 
     type muon_array_t is array(0 to MUON_ARRAY_LENGTH-1) of std_logic_vector(MUON_DATA_WIDTH-1 downto 0);
@@ -139,13 +149,19 @@ package lhc_data_pkg is
         htm : std_logic_vector(HTM_DATA_WIDTH-1 downto 0);
         etmhf : std_logic_vector(ETMHF_DATA_WIDTH-1 downto 0);
         htmhf : std_logic_vector(HTMHF_DATA_WIDTH-1 downto 0);
-        zdc10g_0 : std_logic_vector(ZDC10G_0_WIDTH-1 downto 0);
-        zdc10g_1 : std_logic_vector(ZDC10G_1_WIDTH-1 downto 0);
-        zdc10g_2 : std_logic_vector(ZDC10G_2_WIDTH-1 downto 0);
-        zdc10g_3 : std_logic_vector(ZDC10G_3_WIDTH-1 downto 0);
-        zdc10g_4 : std_logic_vector(ZDC10G_4_WIDTH-1 downto 0);
-        zdc10g_5 : std_logic_vector(ZDC10G_5_WIDTH-1 downto 0);
+        link_11_fr_0_data : std_logic_vector(LINK_11_FR_0_WIDTH-1 downto 0);
+        link_11_fr_1_data : std_logic_vector(LINK_11_FR_1_WIDTH-1 downto 0);
+        link_11_fr_2_data : std_logic_vector(LINK_11_FR_2_WIDTH-1 downto 0);
+        link_11_fr_3_data : std_logic_vector(LINK_11_FR_3_WIDTH-1 downto 0);
+        link_11_fr_4_data : std_logic_vector(LINK_11_FR_4_WIDTH-1 downto 0);
+        link_11_fr_5_data : std_logic_vector(LINK_11_FR_5_WIDTH-1 downto 0);
         external_conditions : std_logic_vector(EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto 0);
+        ZDC5G_0 : std_logic_vector(ZDC5G_0_WIDTH-1 downto 0);
+        ZDC5G_1 : std_logic_vector(ZDC5G_1_WIDTH-1 downto 0);
+        ZDC5G_2 : std_logic_vector(ZDC5G_2_WIDTH-1 downto 0);
+        ZDC5G_3 : std_logic_vector(ZDC5G_3_WIDTH-1 downto 0);
+        ZDC5G_4 : std_logic_vector(ZDC5G_4_WIDTH-1 downto 0);
+        ZDC5G_5 : std_logic_vector(ZDC5G_5_WIDTH-1 downto 0);
     end record;
 
     constant LHC_DATA_NULL : lhc_data_t :=
@@ -160,16 +176,22 @@ package lhc_data_pkg is
             htm => (others=>'0'),
             etmhf => (others=>'0'),
             htmhf => (others=>'0'),
-            zdc10g_0 => (others=>'0'),
-            zdc10g_1 => (others=>'0'),
-            zdc10g_2 => (others=>'0'),
-            zdc10g_3 => (others=>'0'),
-            zdc10g_4 => (others=>'0'),
-            zdc10g_5 => (others=>'0'),
-            external_conditions => (others=>'0')
+            link_11_fr_0_data => (others=>'0'),
+            link_11_fr_1_data => (others=>'0'),
+            link_11_fr_2_data => (others=>'0'),
+            link_11_fr_3_data => (others=>'0'),
+            link_11_fr_4_data => (others=>'0'),
+            link_11_fr_5_data => (others=>'0'),
+            external_conditions => (others=>'0'),
+            ZDC5G_0 => (others=>'0'),
+            ZDC5G_1 => (others=>'0'),
+            ZDC5G_2 => (others=>'0'),
+            ZDC5G_3 => (others=>'0'),
+            ZDC5G_4 => (others=>'0'),
+            ZDC5G_5 => (others=>'0')
         );
 
-    constant LHC_DATA_OBJECT_COUNT : integer :=17;
+    constant LHC_DATA_OBJECT_COUNT : integer :=23;
     constant INDEX_MUON : integer := 0;
     constant INDEX_EG : integer := 1;
     constant INDEX_TAU : integer := 2;
@@ -180,13 +202,19 @@ package lhc_data_pkg is
     constant INDEX_HTM : integer := 7;
     constant INDEX_ETMHF : integer := 8;
     constant INDEX_HTMHF : integer := 9;
-    constant INDEX_ZDC10G_0 : integer := 10;
-    constant INDEX_ZDC10G_1 : integer := 11;
-    constant INDEX_ZDC10G_2 : integer := 12;
-    constant INDEX_ZDC10G_3 : integer := 13;
-    constant INDEX_ZDC10G_4 : integer := 14;
-    constant INDEX_ZDC10G_5 : integer := 15;
+    constant INDEX_LINK_11_FR_0 : integer := 10;
+    constant INDEX_LINK_11_FR_1 : integer := 11;
+    constant INDEX_LINK_11_FR_2 : integer := 12;
+    constant INDEX_LINK_11_FR_3 : integer := 13;
+    constant INDEX_LINK_11_FR_4 : integer := 14;
+    constant INDEX_LINK_11_FR_5 : integer := 15;
     constant INDEX_EXTERNAL_CONDITIONS : integer := 16;
+    constant INDEX_ZDC5G_0 : integer := 17;
+    constant INDEX_ZDC5G_1 : integer := 18;
+    constant INDEX_ZDC5G_2 : integer := 19;
+    constant INDEX_ZDC5G_3 : integer := 20;
+    constant INDEX_ZDC5G_4 : integer := 21;
+    constant INDEX_ZDC5G_5 : integer := 22;
     type lhc_data_slv_property_t is array (0 to LHC_DATA_OBJECT_COUNT-1) of natural;
 
     constant LHC_DATA_SLV_OBJECT_WIDTH : lhc_data_slv_property_t :=
@@ -201,10 +229,13 @@ package lhc_data_pkg is
             HTM_DATA_WIDTH,
             ETMHF_DATA_WIDTH,
             HTMHF_DATA_WIDTH,
-            ZDC10G_0_WIDTH, ZDC10G_1_WIDTH,
-            ZDC10G_2_WIDTH, ZDC10G_3_WIDTH,
-            ZDC10G_4_WIDTH, ZDC10G_5_WIDTH,
-            EXTERNAL_CONDITIONS_DATA_WIDTH
+            LINK_11_FR_0_WIDTH, LINK_11_FR_1_WIDTH,
+            LINK_11_FR_2_WIDTH, LINK_11_FR_3_WIDTH,
+            LINK_11_FR_4_WIDTH, LINK_11_FR_5_WIDTH,
+            EXTERNAL_CONDITIONS_DATA_WIDTH,
+            ZDC5G_0_WIDTH, ZDC5G_1_WIDTH,
+            ZDC5G_2_WIDTH, ZDC5G_3_WIDTH,
+            ZDC5G_4_WIDTH, ZDC5G_5_WIDTH
         );
 
     constant LHC_DATA_SLV_START_INDICES : lhc_data_slv_property_t; -- see body
@@ -259,21 +290,35 @@ package body lhc_data_pkg is
         ret_value(index + HTMHF_DATA_WIDTH-1 downto index) := data_in.htmhf;
         index := index + HTMHF_DATA_WIDTH;
 
-        ret_value(index + ZDC10G_0_WIDTH-1 downto index) := data_in.zdc10g_0;
-        index := index + ZDC10G_0_WIDTH;
-        ret_value(index + ZDC10G_1_WIDTH-1 downto index) := data_in.zdc10g_1;
-        index := index + ZDC10G_1_WIDTH;
-        ret_value(index + ZDC10G_2_WIDTH-1 downto index) := data_in.zdc10g_2;
-        index := index + ZDC10G_2_WIDTH;
-        ret_value(index + ZDC10G_3_WIDTH-1 downto index) := data_in.zdc10g_3;
-        index := index + ZDC10G_3_WIDTH;
-        ret_value(index + ZDC10G_4_WIDTH-1 downto index) := data_in.zdc10g_4;
-        index := index + ZDC10G_4_WIDTH;
-        ret_value(index + ZDC10G_5_WIDTH-1 downto index) := data_in.zdc10g_5;
-        index := index + ZDC10G_5_WIDTH;
+        ret_value(index + LINK_11_FR_0_WIDTH-1 downto index) := data_in.link_11_fr_0_data;
+        index := index + LINK_11_FR_0_WIDTH;
+        ret_value(index + LINK_11_FR_1_WIDTH-1 downto index) := data_in.link_11_fr_1_data;
+        index := index + LINK_11_FR_1_WIDTH;
+        ret_value(index + LINK_11_FR_2_WIDTH-1 downto index) := data_in.link_11_fr_2_data;
+        index := index + LINK_11_FR_2_WIDTH;
+        ret_value(index + LINK_11_FR_3_WIDTH-1 downto index) := data_in.link_11_fr_3_data;
+        index := index + LINK_11_FR_3_WIDTH;
+        ret_value(index + LINK_11_FR_4_WIDTH-1 downto index) := data_in.link_11_fr_4_data;
+        index := index + LINK_11_FR_4_WIDTH;
+        ret_value(index + LINK_11_FR_5_WIDTH-1 downto index) := data_in.link_11_fr_5_data;
+        index := index + LINK_11_FR_5_WIDTH;
 
         ret_value(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index) := data_in.external_conditions;
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
+        
+        ret_value(index + ZDC5G_0_WIDTH-1 downto index) := data_in.ZDC5G_0;
+        index := index + ZDC5G_0_WIDTH;
+        ret_value(index + ZDC5G_1_WIDTH-1 downto index) := data_in.ZDC5G_1;
+        index := index + ZDC5G_1_WIDTH;
+        ret_value(index + ZDC5G_2_WIDTH-1 downto index) := data_in.ZDC5G_2;
+        index := index + ZDC5G_2_WIDTH;
+        ret_value(index + ZDC5G_3_WIDTH-1 downto index) := data_in.ZDC5G_3;
+        index := index + ZDC5G_3_WIDTH;
+        ret_value(index + ZDC5G_4_WIDTH-1 downto index) := data_in.ZDC5G_4;
+        index := index + ZDC5G_4_WIDTH;
+        ret_value(index + ZDC5G_5_WIDTH-1 downto index) := data_in.ZDC5G_5;
+        index := index + ZDC5G_5_WIDTH;
+
         return ret_value;
     end function;
 
@@ -316,21 +361,35 @@ package body lhc_data_pkg is
         ret_value.htmhf := data_in(index + HTMHF_DATA_WIDTH-1 downto index);
         index := index + HTMHF_DATA_WIDTH;
 
-        ret_value.zdc10g_0 := data_in(index + ZDC10G_0_WIDTH-1 downto index);
-        index := index + ZDC10G_0_WIDTH;
-        ret_value.zdc10g_1 := data_in(index + ZDC10G_1_WIDTH-1 downto index);
-        index := index + ZDC10G_1_WIDTH;
-        ret_value.zdc10g_2 := data_in(index + ZDC10G_2_WIDTH-1 downto index);
-        index := index + ZDC10G_2_WIDTH;
-        ret_value.zdc10g_3 := data_in(index + ZDC10G_3_WIDTH-1 downto index);
-        index := index + ZDC10G_3_WIDTH;
-        ret_value.zdc10g_4 := data_in(index + ZDC10G_4_WIDTH-1 downto index);
-        index := index + ZDC10G_4_WIDTH;
-        ret_value.zdc10g_5 := data_in(index + ZDC10G_5_WIDTH-1 downto index);
-        index := index + ZDC10G_5_WIDTH;
+        ret_value.link_11_fr_0_data := data_in(index + LINK_11_FR_0_WIDTH-1 downto index);
+        index := index + LINK_11_FR_0_WIDTH;
+        ret_value.link_11_fr_1_data := data_in(index + LINK_11_FR_1_WIDTH-1 downto index);
+        index := index + LINK_11_FR_1_WIDTH;
+        ret_value.link_11_fr_2_data := data_in(index + LINK_11_FR_2_WIDTH-1 downto index);
+        index := index + LINK_11_FR_2_WIDTH;
+        ret_value.link_11_fr_3_data := data_in(index + LINK_11_FR_3_WIDTH-1 downto index);
+        index := index + LINK_11_FR_3_WIDTH;
+        ret_value.link_11_fr_4_data := data_in(index + LINK_11_FR_4_WIDTH-1 downto index);
+        index := index + LINK_11_FR_4_WIDTH;
+        ret_value.link_11_fr_5_data := data_in(index + LINK_11_FR_5_WIDTH-1 downto index);
+        index := index + LINK_11_FR_5_WIDTH;
 
         ret_value.external_conditions := data_in(index + EXTERNAL_CONDITIONS_DATA_WIDTH-1 downto index);
         index := index + EXTERNAL_CONDITIONS_DATA_WIDTH;
+
+        ret_value.ZDC5G_0 := data_in(index + ZDC5G_0_WIDTH-1 downto index);
+        index := index + ZDC5G_0_WIDTH;
+        ret_value.ZDC5G_1 := data_in(index + ZDC5G_1_WIDTH-1 downto index);
+        index := index + ZDC5G_1_WIDTH;
+        ret_value.ZDC5G_2 := data_in(index + ZDC5G_2_WIDTH-1 downto index);
+        index := index + ZDC5G_2_WIDTH;
+        ret_value.ZDC5G_3 := data_in(index + ZDC5G_3_WIDTH-1 downto index);
+        index := index + ZDC5G_3_WIDTH;
+        ret_value.ZDC5G_4 := data_in(index + ZDC5G_4_WIDTH-1 downto index);
+        index := index + ZDC5G_4_WIDTH;
+        ret_value.ZDC5G_5 := data_in(index + ZDC5G_5_WIDTH-1 downto index);
+        index := index + ZDC5G_5_WIDTH;
+
         return ret_value;
     end function;