+++ firmware/hdl/payload/frame/lmp.vhd
@@ -2,6 +2,7 @@
 -- Lane mapping process.
 
 -- Version-history:
+-- HB 2023-08-25: zdc5g on link 71, no zdc10g.
 -- HB 2022-10-08: zdc10g on link 11.
 -- HB 2022-09-06: cleaned up.
 -- HB 2106-05-31: lane mapping for all frames of calo links (for extended test-vector-file structure - see lhc_data_pkg.vhd)
@@ -12,7 +13,6 @@ use IEEE.std_logic_1164.all;
 
 use work.gt_mp7_core_pkg.all;
 use work.lhc_data_pkg.all;
-use work.gtl_pkg.all;
 
 entity lmp is
     generic
@@ -25,7 +25,6 @@ entity lmp is
         demux_data_valid_i : in demux_lanes_data_objects_array_valid_t(NR_LANES-1 downto 0);
         lhc_data_o : out lhc_data_t;
         lhc_data_valid_o : out std_logic
---         zdc5g : out zdc_array
     );
 end;
 
@@ -38,10 +37,10 @@ architecture arch of lmp is
 --     constant OFFSET_JET_LANES : natural := 6;
 --     constant OFFSET_TAU_LANES : natural := 8;
 --     constant OFFSET_ESUMS_LANES : natural := 10;
--- -- HB 2106-05-31: proposal for memory structure with all frames of calo links for extended test-vector-file structure (see lhc_data_pkg.vhd)
---     constant OFFSET_ZDC10G_LANES : natural := 11;
 -- -- BR 2015-05-01: added external-conditions.
 --     constant OFFSET_EXT_COND_LANES : natural := 12;
+--     constant OFFSET_LINK_11_LANES : natural := 11;
+--     constant OFFSET_ZDC5G_LANES : natural := 71;
 
 begin
 
@@ -121,12 +120,19 @@ begin
     lhc_data_o.etmhf <= demux_data_i(OFFSET_ESUMS_LANES)(4);
     lhc_data_o.htmhf <= demux_data_i(OFFSET_ESUMS_LANES)(5);
 
-    lhc_data_o.zdc10g_0 <= demux_data_i(OFFSET_ZDC10G_LANES)(0);
-    lhc_data_o.zdc10g_1 <= demux_data_i(OFFSET_ZDC10G_LANES)(1);
-    lhc_data_o.zdc10g_2 <= demux_data_i(OFFSET_ZDC10G_LANES)(2);
-    lhc_data_o.zdc10g_3 <= demux_data_i(OFFSET_ZDC10G_LANES)(3);
-    lhc_data_o.zdc10g_4 <= demux_data_i(OFFSET_ZDC10G_LANES)(4);
-    lhc_data_o.zdc10g_5 <= demux_data_i(OFFSET_ZDC10G_LANES)(5);
+    lhc_data_o.link_11_fr_0_data <= demux_data_i(OFFSET_LINK_11_LANES)(0);
+    lhc_data_o.link_11_fr_1_data <= demux_data_i(OFFSET_LINK_11_LANES)(1);
+    lhc_data_o.link_11_fr_2_data <= demux_data_i(OFFSET_LINK_11_LANES)(2);
+    lhc_data_o.link_11_fr_3_data <= demux_data_i(OFFSET_LINK_11_LANES)(3);
+    lhc_data_o.link_11_fr_4_data <= demux_data_i(OFFSET_LINK_11_LANES)(4);
+    lhc_data_o.link_11_fr_5_data <= demux_data_i(OFFSET_LINK_11_LANES)(5);
+
+    lhc_data_o.zdc5g_0 <= demux_data_i(OFFSET_ZDC5G_LANES)(0);
+    lhc_data_o.zdc5g_1 <= demux_data_i(OFFSET_ZDC5G_LANES)(1);
+    lhc_data_o.zdc5g_2 <= demux_data_i(OFFSET_ZDC5G_LANES)(2);
+    lhc_data_o.zdc5g_3 <= demux_data_i(OFFSET_ZDC5G_LANES)(3);
+    lhc_data_o.zdc5g_4 <= demux_data_i(OFFSET_ZDC5G_LANES)(4);
+    lhc_data_o.zdc5g_5 <= demux_data_i(OFFSET_ZDC5G_LANES)(5);
 
 -- HEPHY 2015-05-01: added external-conditions.
     lhc_data_o.external_conditions(31 downto 0) <= demux_data_i(OFFSET_EXT_COND_LANES+0)(0);
@@ -138,12 +144,4 @@ begin
     lhc_data_o.external_conditions(223 downto 192) <= demux_data_i(OFFSET_EXT_COND_LANES+3)(0);
     lhc_data_o.external_conditions(255 downto 224) <= demux_data_i(OFFSET_EXT_COND_LANES+3)(1);
 
---     zdc5g(0) <= demux_data_i(ZDC5G_LANE_NR)(0);
---     zdc5g(1) <= demux_data_i(ZDC5G_LANE_NR)(1);
---     zdc5g(2) <= demux_data_i(ZDC5G_LANE_NR)(2);
---     zdc5g(3) <= demux_data_i(ZDC5G_LANE_NR)(3);
---     zdc5g(4) <= demux_data_i(ZDC5G_LANE_NR)(4);
---     zdc5g(5) <= demux_data_i(ZDC5G_LANE_NR)(5);
-
 end architecture;
-